# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Educational & Career Development Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  points: number;
  streak?: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  role?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface Job {
  id: number;
  title: string;
  company: string;
  description: string;
  type: string;
  level: string;
  /** @nullable */
  location?: string | null;
  isRemote: boolean;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  status: string;
  passScore: number;
  applicationCount?: number;
  createdAt: string;
}

export interface JobInput {
  title: string;
  company: string;
  description: string;
  type: string;
  level: string;
  location?: string;
  isRemote: boolean;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  passScore: number;
}

export interface JobUpdate {
  title?: string;
  company?: string;
  description?: string;
  type?: string;
  level?: string;
  /** @nullable */
  location?: string | null;
  isRemote?: boolean;
  /** @nullable */
  salaryMin?: number | null;
  /** @nullable */
  salaryMax?: number | null;
  status?: string;
  passScore?: number;
}

export interface JobStats {
  totalJobs: number;
  openJobs: number;
  totalApplications: number;
  filledPositions: number;
}

export interface ScreeningQuestion {
  id: number;
  jobId: number;
  question: string;
  options: string[];
  correctIndex: number;
  order: number;
}

export interface ScreeningQuestionInput {
  question: string;
  options: string[];
  correctIndex: number;
  order?: number;
}

export interface Application {
  id: number;
  jobId: number;
  /** @nullable */
  jobTitle?: string | null;
  applicantName: string;
  applicantEmail: string;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  status: string;
  /** @nullable */
  screeningScore?: number | null;
  /** @nullable */
  screeningPassed?: boolean | null;
  /** @nullable */
  userId?: number | null;
  createdAt: string;
}

export interface ApplicationInput {
  jobId: number;
  applicantName: string;
  applicantEmail: string;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  userId?: number | null;
}

export interface ApplicationStatusUpdate {
  status: string;
}

export interface ScreeningSubmission {
  answers: number[];
}

export interface ScreeningResult {
  score: number;
  passed: boolean;
  total: number;
  message?: string;
}

export interface Workshop {
  id: number;
  title: string;
  description: string;
  date: string;
  duration: number;
  instructor: string;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  status: string;
  capacity: number;
  enrolledCount: number;
  passScore: number;
  createdAt: string;
}

export interface WorkshopInput {
  title: string;
  description: string;
  date: string;
  duration: number;
  instructor: string;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  capacity: number;
  passScore: number;
}

export interface WorkshopUpdate {
  title?: string;
  description?: string;
  date?: string;
  duration?: number;
  instructor?: string;
  tags?: string[];
  /** @nullable */
  imageUrl?: string | null;
  status?: string;
  capacity?: number;
  passScore?: number;
}

export interface EnrollInput {
  userId: number;
  userName: string;
  userEmail: string;
}

export interface Enrollment {
  id: number;
  workshopId: number;
  userId: number;
  userName: string;
  userEmail: string;
  createdAt: string;
}

export interface ExamQuestion {
  id: number;
  workshopId: number;
  question: string;
  options: string[];
  order: number;
}

export interface Exam {
  workshopId: number;
  workshopTitle: string;
  passScore: number;
  /** @nullable */
  timeLimitMinutes?: number | null;
  questions: ExamQuestion[];
}

export interface ExamQuestionInput {
  question: string;
  options: string[];
  correctIndex: number;
  order?: number;
}

export interface ExamSubmission {
  userId: number;
  answers: number[];
}

export interface ExamResult {
  score: number;
  passed: boolean;
  total: number;
  certificateIssued: boolean;
  /** @nullable */
  certificateId?: number | null;
  message?: string;
}

export interface Certificate {
  id: number;
  userId: number;
  userName: string;
  workshopId?: number;
  workshopTitle: string;
  issuedAt: string;
  score: number;
  certificateNumber?: string;
}

export interface Track {
  id: number;
  slug: string;
  title: string;
  description: string;
  category: string;
  level: string;
  /** @nullable */
  iconUrl?: string | null;
  moduleCount: number;
  estimatedHours: number;
  enrolledCount?: number;
}

export interface TrackModule {
  id: number;
  trackId: number;
  title: string;
  description: string;
  type: string;
  /** @nullable */
  content?: string | null;
  order: number;
  estimatedMinutes?: number;
}

export interface TrackDetail {
  id: number;
  slug: string;
  title: string;
  description: string;
  category: string;
  level: string;
  /** @nullable */
  iconUrl?: string | null;
  moduleCount: number;
  estimatedHours: number;
  enrolledCount?: number;
  modules: TrackModule[];
}

export interface TrackProgress {
  userId: number;
  trackSlug: string;
  completedModules: number[];
  totalModules: number;
  percentComplete: number;
  points: number;
}

export interface ProgressUpdate {
  userId: number;
  moduleId: number;
  completed: boolean;
}

export interface LeaderboardEntry {
  rank: number;
  userId: number;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  points: number;
  streak: number;
  certificateCount: number;
  completedTracks: number;
}

export interface PlatformStats {
  studentsTrained: number;
  certificatesIssued: number;
  jobsFilled: number;
  workshopsHeld: number;
  activeJobs: number;
}

export interface AdminStats {
  totalUsers: number;
  totalJobs: number;
  openJobs: number;
  totalApplications: number;
  pendingApplications: number;
  totalWorkshops: number;
  upcomingWorkshops: number;
  totalCertificates: number;
  recentApplications: Application[];
}

export interface MockInterviewSession {
  id: number;
  userId: number;
  track: string;
  title?: string;
  createdAt: string;
}

export interface MockInterviewSessionInput {
  userId: number;
  track: string;
  title?: string;
}

export interface MockInterviewMessage {
  sessionId: number;
  message: string;
  role: string;
}

export interface MockInterviewResponse {
  reply: string;
  feedback: string;
  /** @nullable */
  score?: number | null;
}

export type ListJobsParams = {
type?: string;
level?: string;
search?: string;
remote?: string;
};

export type ListApplicationsParams = {
jobId?: number;
status?: string;
};

export type ListWorkshopsParams = {
status?: string;
};

export type GetLeaderboardParams = {
limit?: number;
};