# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Educational & Career Development Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    patch:
      operationId: updateJob
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/screening-questions:
    get:
      operationId: getJobScreeningQuestions
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Screening quiz questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScreeningQuestion"
    post:
      operationId: addJobScreeningQuestion
      tags: [jobs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningQuestionInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningQuestion"

  # ── Applications ──────────────────────────────────────────────────────
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplicationStatus
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/screening:
    post:
      operationId: submitScreening
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreeningSubmission"
      responses:
        "200":
          description: Screening result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreeningResult"

  # ── Workshops ─────────────────────────────────────────────────────────
...
  /certificates/{id}:
    get:
      operationId: getCertificate
      tags: [certificates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # ── Learning Tracks ───────────────────────────────────────────────────
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, points, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
        points:
          type: integer
        streak:
          type: integer
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Job:
      type: object
      required: [id, title, company, description, type, level, isRemote, status, passScore, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        description:
          type: string
        type:
          type: string
        level:
          type: string
        location:
          type: ["string", "null"]
        isRemote:
          type: boolean
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        status:
          type: string
        passScore:
          type: integer
        applicationCount:
          type: integer
        createdAt:
          type: string
    JobInput:
      type: object
      required: [title, company, description, type, level, isRemote, passScore]
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: string
        type:
          type: string
        level:
          type: string
        location:
          type: string
        isRemote:
          type: boolean
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        passScore:
          type: integer
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        company:
          type: string
        description:
          type: string
        type:
          type: string
        level:
          type: string
        location:
          type: ["string", "null"]
        isRemote:
          type: boolean
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        status:
          type: string
        passScore:
          type: integer
...
    ScreeningQuestion:
      type: object
      required: [id, jobId, question, options, correctIndex, order]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        order:
          type: integer
    ScreeningQuestionInput:
      type: object
      required: [question, options, correctIndex]
      properties:
        question:
          type: string
        options:
          type: array
          items:
            type: string
        correctIndex:
          type: integer
        order:
          type: integer

    # ── Applications ──────────────────────────────
    Application:
      type: object
      required: [id, jobId, applicantName, applicantEmail, status, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        jobTitle:
          type: ["string", "null"]
        applicantName:
          type: string
        applicantEmail:
          type: string
        resumeUrl:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        status:
          type: string
        screeningScore:
          type: ["integer", "null"]
        screeningPassed:
          type: ["boolean", "null"]
        userId:
          type: ["integer", "null"]
        createdAt:
          type: string
    ApplicationInput:
      type: object
      required: [jobId, applicantName, applicantEmail]
      properties:
        jobId:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        resumeUrl:
          type: ["string", "null"]
        coverLetter:
          type: ["string", "null"]
        userId:
          type: ["integer", "null"]
    ApplicationStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
    ScreeningSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: integer
    ScreeningResult:
      type: object
      required: [score, passed, total]
      properties:
        score:
          type: integer
        passed:
          type: boolean
        total:
          type: integer
        message:
          type: string

    # ── Workshops ─────────────────────────────────
...
    Certificate:
      type: object
      required: [id, userId, userName, workshopTitle, issuedAt, score]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        workshopId:
          type: integer
        workshopTitle:
          type: string
        issuedAt:
          type: string
        score:
          type: integer
        certificateNumber:
          type: string

    # ── Learning Tracks ────────────────────────────
...