# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Educational & Career Development Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  points: number;
  streak?: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}